requirements.txt

streamlit
pandas
plotly
openpyxl
facebook-business
requests
python-dotenv